% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parcel_atlas.R
\name{get_parcel_atlas}
\alias{get_parcel_atlas}
\title{Get parcel-based structural connectivity}
\usage{
get_parcel_atlas(cfg)
}
\arguments{
\item{cfg}{a pre-made cfg structure (as list object).}
}
\value{
An .RData file with the suffix .connectivity.RData, which contains the structural connection matrix (connectivity);
an .RData file with the suffix .network_measures.RData, which contains various graph measures for the SC matrix;
a .txt file with the suffix .connectogram.txt, which contains a connectogram that can be viewed on http://mkweb.bcgsc.ca/tableviewer/visualize/ by checking the two size options in step 2A (col with row size, row with col size);
a .tdi.nii.gz file named the same way as the .trk.gz file, which contains a nifti image with track density imaging (TDI) values at each voxel. It is essentially a way of converting the .trk.gz file into voxel space. Higher values indicate higher streamline densities at each voxel.
}
\description{
This function computes parcel-based direct structural connectivity measures using an MNI-registered
brain parcellation and the curated HCP-842 structural connectome template described in Yeh et al., (2018 - NeuroImage)
}
