% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parcel_cons.R
\name{get_parcel_cons}
\alias{get_parcel_cons}
\title{Get parcel disconnection measures/maps}
\usage{
get_parcel_cons(cfg, cores = 1)
}
\arguments{
\item{cfg}{a pre-made cfg structure (as list object).}

\item{cores}{an integer value that indicates how many parallel cores the function should be run on.}
}
\description{
This is a wrapper function that calls a series of functions to create various parcel disconnection measures.
It wraps the 'get_parcel_atlas', 'get_atlas_sspl', 'get_parcel_discon', and 'get_patient_sspl' functions.
}
