% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_patient_sspl.R
\name{get_patient_sspl}
\alias{get_patient_sspl}
\title{Get structural shortest path length (SSPL)}
\usage{
get_patient_sspl(cfg, cores = 1)
}
\arguments{
\item{cfg}{a pre-made cfg structure (as list object).}

\item{cores}{an integer value that indicates how many parallel cores the function should be run on.}
}
\value{
A file with the suffix _SSPL_matrices.RData -- this contains the raw SSPL matrix (sspl_matrix), the lesion-based increases in SSPLs relative to the atlas (both direct and indirect disconnections; delta_ssplL_matrix), and the indirect-only delta matrix (idc_matrix);
a file with the suffix _delta_SSPL.node -- this contains the node information for loading the delta SSPL matrix into an external viewer. Node sizes are proportional to the total SSPL increases for each node (relative to the max increase);
a file with the suffix _delta_SSPL.edge -- this contains the edge information for loading the delta SSPL matrix into an external viewer;
a file with the suffix _indirect_SDC.node -- same as (3), but for the indirect SDC matrix;
a file with the suffix _indirect_SDC.edge -- same as (4), but for the indirect SDC matrix.

Note that if the lesion leads to the SSPL between a region pair to be undefined (i.e. = Inf), then this value is replaced
by a value equal to 1 plus the maximum non-Inf value in the atlas SSPL matrix.
}
\description{
This function computes increases in the structural shortest path length (SSPL) between brain regions
associated with focal lesions. See Griffis et al. (2019, NeuroImage) for more details.
}
