% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tract_discon.R
\name{get_tract_discon}
\alias{get_tract_discon}
\title{Get tract-based disconnection}
\usage{
get_tract_discon(cfg, cores = 1)
}
\arguments{
\item{cfg}{a pre-made cfg structure (as list object).}

\item{cores}{an integer value that indicates how many parallel cores the function should be run on.}
}
\value{
A .csv file giving the percent of streamlines in each tract that were disconnected by the lesions (column "Discon")
and the associated tract names (column "Tract").
}
\description{
This function computes tract-based disconnection measures using an MNI-registered lesion and the tract segmentations
obtained from the curated HCP-842 tractography atlas as described in Yeh et al., (2018 - NeuroImage).
}
