% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lqt_subject.R
\name{plot_lqt_subject}
\alias{plot_lqt_subject}
\title{Plot subject-level summaries}
\usage{
plot_lqt_subject(cfg, subject = 1, type = NULL)
}
\arguments{
\item{cfg}{a pre-made cfg structure (as list object).}

\item{subject}{either a string giving a subject ID, or an integer giving the subject index.}

\item{type}{a string specifying the desired plot type. Options are "parcel.damage", which gives a barplot of percent parcel damage;
"tract.discon", which gives a barplot of percent tract-level disconnection;
"parcel.discon", which gives a barplot and brain network plots showing parcel-level disconnection;
and "parcel.sspl", which gives a barplot and brain network plots showing parcel-level structural shortest path length increases.}
}
\value{
A plot object visualizing either  parcel damage (type = 'parcel.damage'),
tract disconnection (type = 'tract.discon'), i
nter/intra-parcel structural disconnection (type = 'parcel.discon'),
or changes in inter-parcel structural shortest path lengths (type = 'parcel.sspl').
}
\description{
This function constructs subject-level visualizations of lesion-based damage and disconnection
}
