% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_subject_summaries.R
\name{plot_subject_summaries}
\alias{plot_subject_summaries}
\title{Plot subject-level summaries}
\usage{
plot_subject_summaries(cfg, type = c(""))
}
\arguments{
\item{cfg}{a pre-made cfg structure (as list object).}
}
\value{
A data.frame giving the percent of streamlines in each tract that were disconnected by the lesions (column "Discon")
and the associated tract names (column "Tract").
}
\description{
This function
}
