% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_data.R
\name{compile_data}
\alias{compile_data}
\title{Compile datasets}
\usage{
compile_data(out_path)
}
\arguments{
\item{out_path}{a string specifying the output directory to which LQT files were saved.}
}
\value{
An .RData file with the suffix .connectivity.RData. This contains the structural disconnection matrix (connectivity);
an .RData file with the suffix .network_measures.RData, which contains various graph measures for the SC matrix;
an .RData file with the suffix _percent_parcel_mats.RData. This file contains a disconnection adjacency matrix (pct_sdc_matrix) and a spared connection adjacency matrix (pct_spared_sc_matrix);
}
\description{
This function takes completed LQT files and compiles analysis-ready datasets
}
