% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_data.R
\name{compile_data}
\alias{compile_data}
\title{Compile datasets}
\usage{
compile_data(cfg, cores = 1)
}
\arguments{
\item{cfg}{a pre-made cfg structure (as list object).}

\item{cores}{an integer value that indicates how many parallel cores the function should be run on.}
}
\value{
A list object containing several analysis-ready datasets, in which rows represent patients and columns represent lesion metrics:
"parc.damage", which gives percent regional parcel damage across subjects;
"tract.discon", which gives percent tract disconnection across subjects;
"net.discon", which gives the percent disconnection for each parcel group or network across its edges;
"parc.discon", which gives the percent disconnection for each individual parcel across its edges;
"net2net.discon", which gives the percent disconnection of each pairwise network-to-network edge;
"parc2parc.discon", which gives the percent disconnection of each pairwise parcel-to-parcel edge.
}
\description{
This function takes completed LQT files and compiles analysis-ready datasets
}
