% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parcel_damage.R
\name{get_parcel_damage}
\alias{get_parcel_damage}
\title{Get parcel damage}
\usage{
get_parcel_damage(cfg, cores = 1)
}
\arguments{
\item{cfg}{a pre-made cfg structure (as list object).}

\item{cores}{an integer value that indicates how many parallel cores the function should be run on.}
}
\value{
A list containing
1) damage.map (a nifti image in which voxel values correspond to the % of the associated parcel damaged by lesion),
and 2) damage.vec (a vector with % damage values for each parcel in the brain parcellation).
}
\description{
This function uses an MNI-registered lesion file and an MNI-registered brain parcellation
to estimate the amount of damage sustained by each brain region.
}
