% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_subject_summaries.R
\name{plot_subject_summaries}
\alias{plot_subject_summaries}
\title{Plot subject-level summaries}
\usage{
plot_subject_summaries(cfg, patient = 1)
}
\arguments{
\item{cfg}{a pre-made cfg structure (as list object).}

\item{patient}{either a string giving a patient ID, or an integer giving the patient index.}
}
\value{
A list object containing plots of parcel damage ('parcel.damage'),
tract disconnection ('tract.discon'), inter/intra-parcel structural disconnection ('parcel.discon'),
and changes in inter-parcel structural shortest path lengths ('parcel.sspl').
}
\description{
This function
}
